chore: 종료 시간이 시작 시간보다 클 경우 toast 띄우는 로직 설정
