+++ features/record/utils/index.ts
@@ -1,4 +1,3 @@
-export * from './add-minutes';
-export * from './convert-to-24-hour-format';
 export * from './highlight-text';
 export * from './remove-special-symbols';
+export * from './time';
