+++ features/record/utils/time.ts
@@ -1,5 +1,23 @@
 import { AmpmType, HourType, MinuteType } from '../types';
 
+export const convertTo24HourFormat = (pickerValue: {
+  ampm: AmpmType;
+  hour: HourType;
+  minute: MinuteType;
+}) => {
+  let hour24 = parseInt(pickerValue.hour, 10);
+
+  if (pickerValue.ampm === '오후' && hour24 !== 12) {
+    hour24 += 12;
+  }
+
+  if (pickerValue.ampm === '오전' && hour24 === 12) {
+    hour24 = 0;
+  }
+
+  return `${hour24.toString().padStart(2, '0')}:${pickerValue.minute}`;
+};
+
 export const addMinutes = (
   pickerValue: {
     ampm: AmpmType;
@@ -29,4 +47,14 @@ export const addMinutes = (
   const newMinute = date.getMinutes().toString().padStart(2, '0');
 
   return `${newHour}:${newMinute}`;
+};
+
+export const compareTime = (time1: string, time2: string) => {
+  const [hour1, minute1] = time1.split(':').map(Number);
+  const [hour2, minute2] = time2.split(':').map(Number);
+
+  const totalMinutes1 = hour1 * 60 + minute1;
+  const totalMinutes2 = hour2 * 60 + minute2;
+
+  return totalMinutes1 < totalMinutes2;
 };
